Use all available cores with makepkg
All cores available will now be used when building software and
compressing packages. The Arch Wiki says it may cause issues with
some packages, but testing have so far revealed none. It could
always be made optional if need be.

To better handle these additions the template module is now used.
Earlier reasoning behind using the lineinfile module was possibly
having less issues if yaourt adds features. Both makepkg and
yaourt seem quite stable now days; probably won't be that often.

Now instals yaourtrc, even if yaourt is not installed. It's a small
file, deal with it. If yaourt is installed and overwrites the
existing file, ansible quickly restores it in the following task.

Ignore errors with makepkg. That custom module has its issues.
It is also possible the user have entered incorrect package names,
the package has been removed from aur, or there is just a build
error. There is no need to stop the script because of this.